+++ desktop/appcontainer/AppContainer.js
@@ -144,7 +144,7 @@ function globalHotKeys(model) {
 const idlePanelHost = hoistCmp.factory({
     displayName: 'IdlePanel',
     render() {
-        const content = XH.appSpec.idlePanelClass ?? idlePanel;
+        const content = XH.appSpec.idlePanel ?? idlePanel;
         return elementFromContent(content, {onReactivate: () => XH.reloadApp()});
     }
 });